requirements.txt
email_validator==2.2.0
fastapi==0.116.0
passlib[bcrypt]
psycopg2-binary==2.9.10
python-jose[cryptography]
python-multipart==0.0.20
SQLAlchemy==2.0.41
uvicorn==0.35.0
